Various optimizations
Changed a lot of the core of this application - the way switches and
locks are stored in database has changed to an "Associative array" which
in javascript is really just a general object... also moved some
functionality into separate helper methods. Also made switch retrieval
much more efficient and added automatic polling so that switches will be
polled automatically so if you change the status outside of the app the
switch should update within 5 seconds.